improve-tests

+++ infector/infect_victim.sh
@@ -5,10 +5,20 @@
 
 # please build the //infector package before running this script.
 # run with infector/infect_victim.sh [method] [victim]
-# for example: bazel build //infector/... --config=gcc_x86_64 && infector/infect_victim.sh text_padding bazel-bin/infector/victim
+# for example: bazel build //infector/... && infector/infect_victim.sh text_padding bazel-bin/infector/victim
 
 set -e
 set -o pipefail
+
+if [[ -z $1 || -z $2 ]]; then
+        echo "Usage infector/infect_victim.sh [method] [victim]"
+        exit 0
+fi
+
+if [[ "$1" != "text_padding" && "$1" != "reverse_text" && "$1" != "pt_note" ]]; then
+        echo "Only text_padding reverse_text pt_note are supported methods. "
+        exit 0
+fi
 
 rm -f /tmp/victim
 # extract virus's text segment
@@ -18,7 +28,7 @@ chmod 700 bazel-bin/infector/test_parasite \
 # copy victim to tmp to be infected
 cp $2 /tmp/victim \
       && chmod 700 /tmp/victim
-echo "infect111111"
+echo "infecting"
 bazel-bin/infector/infector /tmp/victim /tmp/parasite_code $1
 echo "infected"
 # run the infected binary. Since most binaries terminate with --help, this is sufficient to
